Refactors product form
